EES-2940 Move InviteContributor to new ReleaseInviteService
