+++ src/GovUk.Education.ExploreEducationStatistics.Admin.Tests/Services/UserManagementServicePermissionTests.cs
@@ -7,15 +7,11 @@ using GovUk.Education.ExploreEducationStatistics.Admin.Services;
 using GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces;
 using GovUk.Education.ExploreEducationStatistics.Common.Services.Interfaces.Security;
 using GovUk.Education.ExploreEducationStatistics.Common.Utils;
-using GovUk.Education.ExploreEducationStatistics.Content.Model;
 using GovUk.Education.ExploreEducationStatistics.Content.Model.Database;
-using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.Configuration;
 using Moq;
 using Xunit;
 using static GovUk.Education.ExploreEducationStatistics.Admin.Security.SecurityPolicies;
 using static GovUk.Education.ExploreEducationStatistics.Admin.Tests.Services.DbUtils;
-using static GovUk.Education.ExploreEducationStatistics.Common.Services.CollectionUtils;
 using static GovUk.Education.ExploreEducationStatistics.Common.Tests.Utils.MockUtils;
 using static GovUk.Education.ExploreEducationStatistics.Common.Tests.Utils.PermissionTestUtils;
 using static Moq.MockBehavior;
@@ -112,41 +108,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Services
         }
 
         [Fact]
-        public async Task InviteContributor()
-        {
-            var release = new Release();
-            var publication = new Publication
-            {
-                Id = Guid.NewGuid(),
-                Releases = ListOf(release),
-            };
-
-            await PolicyCheckBuilder<SecurityPolicies>()
-                .SetupResourceCheckToFailWithMatcher<Tuple<Publication, ReleaseRole>>(
-                    tuple => tuple.Item1.Id == publication.Id && tuple.Item2 == ReleaseRole.Contributor,
-                    CanUpdateSpecificReleaseRole)
-                .AssertForbidden(async userService =>
-                {
-                    var contentDbContextId = Guid.NewGuid().ToString();
-                    await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-                    {
-                        await contentDbContext.AddAsync(publication);
-                        await contentDbContext.SaveChangesAsync();
-                    }
-
-                    await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-                    {
-                        var service = SetupUserManagementService(
-                            contentDbContext: contentDbContext,
-                            userService: userService.Object);
-                        return await service.InviteContributor("test@test.com",
-                            publication.Id,
-                            ListOf(release.Id));
-                    }
-                });
-        }
-
-        [Fact]
         public async Task CancelInvite()
         {
             await PolicyCheckBuilder<SecurityPolicies>()
@@ -174,18 +135,11 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Services
         private static UserManagementService SetupUserManagementService(
             ContentDbContext? contentDbContext = null,
             UsersAndRolesDbContext? usersAndRolesDbContext = null,
-            IPersistenceHelper<ContentDbContext>? contentPersistenceHelper = null,
             IPersistenceHelper<UsersAndRolesDbContext>? usersAndRolesPersistenceHelper = null,
             IEmailTemplateService? emailTemplateService = null,
             IUserRoleService? userRoleService = null,
-            IUserRepository? userRepository = null,
             IUserService? userService = null,
-            IUserInviteRepository? userInviteRepository = null,
-            IUserReleaseInviteRepository? userReleaseInviteRepository = null,
-            IUserReleaseRoleRepository? userReleaseRoleRepository = null,
-            IConfiguration? configuration = null,
-            IEmailService? emailService = null,
-            IHttpContextAccessor? httpContextAccessor = null)
+            IUserInviteRepository? userInviteRepository = null)
         {
             contentDbContext ??= InMemoryApplicationDbContext();
             usersAndRolesDbContext ??= InMemoryUserAndRolesDbContext();
@@ -193,18 +147,11 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Services
             return new UserManagementService(
                 usersAndRolesDbContext,
                 contentDbContext,
-                contentPersistenceHelper ?? new PersistenceHelper<ContentDbContext>(contentDbContext),
                 usersAndRolesPersistenceHelper ?? new PersistenceHelper<UsersAndRolesDbContext>(usersAndRolesDbContext),
                 emailTemplateService ?? new Mock<IEmailTemplateService>(Strict).Object,
                 userRoleService ?? new Mock<IUserRoleService>(Strict).Object,
-                userRepository ?? new UserRepository(contentDbContext),
                 userService ?? AlwaysTrueUserService().Object,
-                userInviteRepository ?? new UserInviteRepository(usersAndRolesDbContext),
-                userReleaseInviteRepository ?? new UserReleaseInviteRepository(contentDbContext),
-                userReleaseRoleRepository ?? new UserReleaseRoleRepository(contentDbContext),
-                configuration ?? CreateMockConfiguration().Object,
-                emailService ?? new Mock<IEmailService>(Strict).Object,
-                httpContextAccessor ?? new Mock<IHttpContextAccessor>(Strict).Object
+                userInviteRepository ?? new UserInviteRepository(usersAndRolesDbContext)
             );
         }
     }
