+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Services/Interfaces/IReleaseInviteService.cs
@@ -1,0 +1,14 @@
+#nullable enable
+using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using GovUk.Education.ExploreEducationStatistics.Common.Model;
+using Microsoft.AspNetCore.Mvc;
+
+namespace GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces
+{
+    public interface IReleaseInviteService
+    {
+        Task<Either<ActionResult, Unit>> InviteContributor(string email, Guid publicationId, List<Guid> releaseIds);
+    }
+}
